back to float16

+++ config/train_token_completion.py
@@ -33,7 +33,7 @@ warmup_iters = 100 # not super necessary potentially
 backend = 'nccl' # 'nccl', 'gloo', etc.
 # system
 device = 'cuda:0' # examples: 'cpu', 'cuda', 'cuda:0', 'cuda:1' etc., or try 'mps' on macbooks
-dtype = 'bfloat16' # 'float32', 'bfloat16', or 'float16', the latter will auto implement a GradScaler
+dtype = 'float16' # 'float32', 'bfloat16', or 'float16', the latter will auto implement a GradScaler
 compile = False # use PyTorch 2.0 to compile the model to be fastersussuss
